jeager - refactoring
