+++ customer/src/main/java/org/oncors/api/CompanyEndpoint.java
@@ -1,7 +1,9 @@
 package org.oncors.api;
 
 import lombok.extern.slf4j.Slf4j;
+import org.oncors.aop.Trace;
 import org.oncors.model.Company;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -16,33 +18,38 @@ import java.util.List;
 @RequestMapping("/companies")
 public class CompanyEndpoint {
 
+    @Trace
     @GetMapping
     @ResponseStatus(HttpStatus.OK)
-    public ResponseEntity<List<Company>> getCompanies() {
+    public ResponseEntity<List<Company>> getCompanies(@RequestHeader HttpHeaders httpHeaders) {
         throw new NotImplementedException();
     }
 
+    @Trace
     @GetMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public ResponseEntity<Company> getCompany(@PathVariable long id) {
+    public ResponseEntity<Company> getCompany(@RequestHeader HttpHeaders httpHeaders, @PathVariable long id) {
         throw new NotImplementedException();
     }
 
+    @Trace
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public ResponseEntity<Company> postCompany(@RequestBody Company company) {
+    public ResponseEntity<Company> postCompany(@RequestHeader HttpHeaders httpHeaders, @RequestBody Company company) {
         throw new NotImplementedException();
     }
 
+    @Trace
     @PutMapping
     @ResponseStatus(HttpStatus.ACCEPTED)
-    public ResponseEntity<Company> putCompany(@PathVariable long id, @RequestBody Company company) {
+    public ResponseEntity<Company> putCompany(@RequestHeader HttpHeaders httpHeaders, @PathVariable long id, @RequestBody Company company) {
         throw new NotImplementedException();
     }
 
+    @Trace
     @DeleteMapping
     @ResponseStatus(HttpStatus.ACCEPTED)
-    public void deleteCompany(@PathVariable long id) {
+    public void deleteCompany(@RequestHeader HttpHeaders httpHeaders, @PathVariable long id) {
         throw new NotImplementedException();
     }
