+++ shop-auth-service/pom.xml
@@ -19,6 +19,9 @@
         <lombok.version>1.18.12</lombok.version>
         <javax.version>3.0-b71</javax.version>
         <swagger.version>2.9.2</swagger.version>
+        <opentracing-api.version>0.32.0</opentracing-api.version>
+        <jaeger-client.version>0.34.0</jaeger-client.version>
+        <trace-utils.version>1.0-SNAPSHOT</trace-utils.version>
     </properties>
 
     <dependencies>
@@ -94,12 +97,18 @@
         <dependency>
             <groupId>io.jaegertracing</groupId>
             <artifactId>jaeger-client</artifactId>
-            <version>0.34.0</version>
+            <version>${jaeger-client.version}</version>
         </dependency>
         <dependency>
             <groupId>io.opentracing</groupId>
             <artifactId>opentracing-api</artifactId>
-            <version>0.31.0</version>
+            <version>${opentracing-api.version}</version>
+            <scope>compile</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.oncors</groupId>
+            <artifactId>trace-utils</artifactId>
+            <version>${trace-utils.version}</version>
             <scope>compile</scope>
         </dependency>
     </dependencies>
