+++ shop-auth-service/src/main/java/com/oncors/api/AuthApi.java
@@ -2,20 +2,24 @@ package com.oncors.api;
 
 import com.oncors.dto.AuthRequest;
 import com.oncors.service.TokenService;
-import io.opentracing.Scope;
 import io.opentracing.Span;
 import io.opentracing.SpanContext;
 import io.opentracing.Tracer;
 import io.opentracing.propagation.Format;
 import io.opentracing.propagation.TextMapExtractAdapter;
+import jdk.jfr.StackTrace;
 import lombok.extern.slf4j.Slf4j;
+import org.oncors.aop.Trace;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.core.token.Token;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.context.request.RequestContextHolder;
 import org.springframework.web.server.ResponseStatusException;
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
@@ -26,11 +30,10 @@ import springfox.documentation.swagger2.annotations.EnableSwagger2;
 public class AuthApi {
     @Autowired
     private TokenService tokenService;
-    @Autowired
-    Tracer tracer;
 
+    @Trace
     @PostMapping(value = "/authenticate")
-    public ResponseEntity<Token> authenticate(@RequestBody AuthRequest authRequest) {
+    public ResponseEntity<Token> authenticate(@RequestHeader HttpHeaders headers, @RequestBody AuthRequest authRequest) {
         final Token jwtToken;
         try {
             jwtToken = tokenService.generateToken(authRequest.getUsername(), authRequest.getPassword());
@@ -40,13 +43,9 @@ public class AuthApi {
         return ResponseEntity.ok(jwtToken);
     }
 
+    @Trace
     @PostMapping(value = "/validate_token")
-    public ResponseEntity<Boolean> validateToken(String jwtToken, @RequestHeader HttpHeaders headers) {
-        SpanContext parentContext = tracer.extract(Format.Builtin.HTTP_HEADERS, new TextMapExtractAdapter(headers.toSingleValueMap()));
-        Span span = tracer.buildSpan("validate-token").asChildOf(parentContext).start();
-
-
-        span.finish();
+    public ResponseEntity<Boolean> validateToken(@RequestHeader HttpHeaders headers, String jwtToken) {
         log.info("\nvalidate token: " + jwtToken + "\nfor sessionId: " +
                 RequestContextHolder.currentRequestAttributes().getSessionId());
 
@@ -57,20 +56,6 @@ public class AuthApi {
             throw new ResponseStatusException(HttpStatus.NOT_ACCEPTABLE, exception.getMessage());
         }
         return ResponseEntity.ok(isTokenValid);
-    }
-
-    //    @PostMapping(value = "/hello")
-//    public String hello(@RequestHeader HttpHeaders headers) {
-//        SpanContext parentContext = tracer.extract(Format.Builtin.HTTP_HEADERS, new TextMapExtractAdapter(headers.toSingleValueMap()));
-//        Span span = tracer.buildSpan("hello-auth").asChildOf(parentContext).start();
-//        return "hello";
-//    }
-    @GetMapping(value = "/hello")
-    public String hello(@RequestHeader HttpHeaders headers) {
-        SpanContext parentContext = tracer.extract(Format.Builtin.HTTP_HEADERS, new TextMapExtractAdapter(headers.toSingleValueMap()));
-        Span span = tracer.buildSpan("auth-hello-span").asChildOf(parentContext).start();
-        span.finish();
-        return "hello";
     }
 
 }
