+++ utils/auth-utils/pom.xml
@@ -3,7 +3,7 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>shop-sandbox</artifactId>
+        <artifactId>utils</artifactId>
         <groupId>org.oncors</groupId>
         <version>1.0-SNAPSHOT</version>
     </parent>
