+++ utils/auth-utils/src/main/java/org/oncors/model/Authority.java
